!1155 [AIKG] adapt dynamic_shape Merge pull request !1155 from zhuyiyang/dyn

+++ aikg/benchmark/aikgbench/dynamic_shape/elemwise/elemwise_cast_var.py
@@ -3,7 +3,7 @@ import torch.nn as nn
 
 
 class Model(nn.Module):
-    def __init__(self, dtype=torch.float32):
+    def __init__(self, dtype=torch.float16):
         super(Model, self).__init__()
         self.dtype = dtype
 
@@ -20,19 +20,19 @@ class Model(nn.Module):
 def get_inputs_dyn_list():
     # Case 1: Small/Medium shapes
     # Shape (256, 1024) represents a batch of 256 samples with 1024 features each
-    inputs1 = torch.randn(256, 1024, dtype=torch.float64)
+    inputs1 = torch.randn(256, 1024, dtype=torch.float32)
     
     # Case 2: Standard large model shapes
     # Shape (1024, 4096) represents a batch of 1024 samples with 4096 features each
-    inputs2 = torch.randn(1024, 4096, dtype=torch.float64)
+    inputs2 = torch.randn(1024, 4096, dtype=torch.float32)
     
     # Case 3: Large shapes
     # Shape (2048, 8192) represents a batch of 2048 samples with 8192 features each
-    inputs3 = torch.randn(2048, 8192, dtype=torch.float64)
+    inputs3 = torch.randn(2048, 8192, dtype=torch.float32)
     
     # Case 4: Non-16-aligned shapes
     # Shape (125, 5120) represents a batch of 125 samples with 5120 features each
-    inputs4 = torch.randn(125, 5120, dtype=torch.float64)
+    inputs4 = torch.randn(125, 5120, dtype=torch.float32)
     
     return [
         [inputs1],  # Case 1 inputs
@@ -44,5 +44,5 @@ def get_inputs_dyn_list():
 
 def get_init_inputs():
     # Parameters for cast
-    dtype = torch.float32  # Target dtype
+    dtype = torch.float16  # Target dtype
     return [dtype]
