+++ aikg/benchmark/aikgbench/dynamic_shape/elemwise/elemwise_swish_001_var.py
@@ -25,19 +25,19 @@ def get_inputs_dyn_list():
     
     # Case 1: 16-aligned batch, 16-aligned hidden
     # Shape (256, 4096) represents a batch of 256 samples with 4096 features each
-    var1 = torch.randn(256, 4096, dtype=torch.bfloat16)
+    var1 = torch.randn(256, 4096, dtype=torch.float32)
     
     # Case 2: Non-16-aligned batch, 16-aligned hidden
     # Shape (125, 5120) represents a batch of 125 samples with 5120 features each
-    var2 = torch.randn(125, 5120, dtype=torch.bfloat16)
+    var2 = torch.randn(125, 5120, dtype=torch.float32)
     
     # Case 3: 16-aligned batch, non-16-aligned hidden
     # Shape (512, 6144) represents a batch of 512 samples with 6144 features each
-    var3 = torch.randn(512, 6144, dtype=torch.bfloat16)
+    var3 = torch.randn(512, 6144, dtype=torch.float32)
     
     # Case 4: Large batch size
     # Shape (1024, 8192) represents a batch of 1024 samples with 8192 features each
-    var4 = torch.randn(1024, 8192, dtype=torch.bfloat16)
+    var4 = torch.randn(1024, 8192, dtype=torch.float32)
     
     return [
         [var1],
